ingest_downloads_to_mixpanel/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud==0.34.0
google-cloud-bigquery==3.4.0
google-cloud-bigquery-storage==2.16.2